project01_route_opt/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
psycopg2-binary==2.9.7
psycopg2==2.9.7
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
Pillow==10.1.0
scikit-learn==1.3.2
tensorflow==2.15.0
opencv-python==4.8.1.78
pandas==2.1.4
numpy==1.24.3
ortools==9.7.2996
geopy==2.4.1
requests==2.31.0
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
